Remove: 필요없는 파일 삭제
